Mandelbrot render resolution can be different from swapchain resolution

+++ libraries/vulkan_renderer/private/pipelines/toScreen/VulkanToScreenPipeline.cpp
@@ -142,7 +142,7 @@ VulkanToScreenPipeline::createGfxPipeline(
     viewport.x = 0.0f;
     viewport.y = static_cast<float>(swapChain.swapChainExtent.height);
     viewport.height = -static_cast<float>(swapChain.swapChainExtent.height);
-    viewport.width = swapChain.swapChainExtent.width;
+    viewport.width = static_cast<float>(swapChain.swapChainExtent.width);
     viewport.minDepth = 0.0f;
     viewport.maxDepth = 1.0f;
